+++ crates/cheatnet/src/execution/mod.rs
@@ -2,6 +2,7 @@
 pub mod cairo1_execution;
 pub mod calls;
 pub mod contract_print;
+pub mod deprecated;
 pub mod entry_point;
 #[allow(clippy::module_name_repetitions)]
 pub mod execution_info;